image-processor-app/requirements.txt
PyQt6>=6.5.0
rawpy>=0.19.0
Pillow>=10.0.0
numpy>=1.21.0
PyInstaller>=6.0.0

# Upload vers Minio
minio>=7.2.0
